+++ src/config.js
@@ -1,8 +1,0 @@
-// src/config.js
-const config = {
-    preview_protocol: 'http',
-    preview_domain: 'localhost',
-    preview_port: 3010  // Domyślna wartość portu to 3010
-};
-
-export default config;