+++ src/config.prod.js
@@ -1,8 +1,0 @@
-// config.js
-const config = {
-    preview_protocol: 'http',
-    preview_domain: 'www.reactstream.com',
-    preview_port: 3010
-};
-
-export default config;